-Cleaned up rendering code

+++ AIPlayer.cpp
@@ -2,7 +2,8 @@
 #include "BoardState.h"
 #include <iostream>
 
-Move AIPlayer::GetMove(Board board, SDL_MouseButtonEvent& mb) {
+Move AIPlayer::GetMove(Board board, SDL_MouseButtonEvent& mb)
+{
 	Move move = Move();
 	move = Minimax({ board, {3, 3, -1}, 0 }, 3, -10000, 10000, true).move;
 	totalDepth += 3;
@@ -44,7 +45,8 @@ BoardState AIPlayer::Minimax(BoardState state, int depth, int alpha, int beta, b
 				maxEval = { boardWithMove, moves[i], eval.evalValue };
 			}
 			alpha = std::max(alpha, eval.evalValue);
-			if (beta <= alpha) {
+			if(beta <= alpha)
+			{
 				prunedNodes++;
 				break;
 			}
@@ -62,11 +64,13 @@ BoardState AIPlayer::Minimax(BoardState state, int depth, int alpha, int beta, b
 			auto eval = Minimax({ boardWithMove, moves[i], 0 }, depth - 1, alpha, beta, !isMax);
 			exploredNodes++;
 
-			if (eval.evalValue < minEval.evalValue) {
+			if(eval.evalValue < minEval.evalValue)
+			{
 				minEval = { boardWithMove, moves[i], eval.evalValue };
 			}
 			beta = std::min(beta, eval.evalValue);
-			if (beta <= alpha) {
+			if(beta <= alpha)
+			{
 				prunedNodes++;
 				break;
 			}
@@ -78,21 +82,24 @@ BoardState AIPlayer::Minimax(BoardState state, int depth, int alpha, int beta, b
 int AIPlayer::GetEvalValue1(Board board)
 {
 	// # max moves - # min moves
-	return GetMoves("White", board).size() - GetMoves("Black", board).size();
+	return static_cast<int>(GetMoves("White", board).size() - GetMoves("Black", board).size());
 }
 
-int AIPlayer::GetEvalValue2(Board board) {
+int AIPlayer::GetEvalValue2(Board board)
+{
 	// # player spaces to move to - # opponent spaces to move to
 	auto playerMoves = GetMoves("White", board);
 	auto opponentMoves = GetMoves("Black", board);
 	int evalValue = 0;
 
-	for (int i = 0; i < playerMoves.size(); i++) {
+	for(int i = 0; i < playerMoves.size(); i++)
+	{
 		Board tempBoard = board;
 		evalValue += tempBoard.MoveStones("White", playerMoves[i].x, playerMoves[i].y, playerMoves[i].direction);
 	}
 
-	for (int i = 0; i < opponentMoves.size(); i++) {
+	for(int i = 0; i < opponentMoves.size(); i++)
+	{
 		Board tempBoard = board;
 		evalValue -= tempBoard.MoveStones("Black", opponentMoves[i].x, opponentMoves[i].y, opponentMoves[i].direction);
 	}
@@ -100,7 +107,8 @@ int AIPlayer::GetEvalValue2(Board board) {
 	return evalValue;
 }
 
-std::vector<Move> AIPlayer::GetMoves(std::string player, Board board) {
+std::vector<Move> AIPlayer::GetMoves(std::string player, Board board)
+{
 	std::vector<Move> possibleMoves;
 	const int directions[] = {
 		85,		// Up
