+++ Button.cpp
@@ -1,4 +1,7 @@
 #include "Button.h"
+#include "RectInt.h"
+#include "Sprite.h"
+#include "SpriteBatch.h"
 
 Button::Button() : rectInt(), sprite(), pressedColor(), isPressed() {}
 
@@ -16,7 +19,7 @@ void Button::OnUp()
 
 void Button::OnRender(SpriteBatch & spriteBatch)
 {
-	spriteBatch.Draw(sprite, Vector2(rectInt.GetX(), rectInt.GetY()), isPressed ? pressedColor : Color::White, 0, Vector2::Zero, Vector2::One, SpriteFlip::None, 0.1f);
+	spriteBatch.Draw(sprite, Vector2(static_cast<float>(rectInt.GetX()), static_cast<float>(rectInt.GetY())), isPressed ? pressedColor : Color::White, 0, Vector2::Zero, Vector2::One, SpriteFlip::None, 0.1f);
 }
 
 bool Button::ContainsPoint(Vector2Int point)
